refactor(cli): what every handler is given has a file of its own
main.ts is back to wiring: a name, a tree, a context, an error map. How the
context is assembled — the environment read, where credentials are kept,
the client built from them — is context.ts's business.
